Dynamic Range Elements
I used angular.js to make the range inputs more dynamic. There might
have been a better way to do this but I used this method in the past and
I didn't want to waste time finding another way. The js file i used is
minimized so the impact on load time should be minimal.